src/main/java/mg/rojo/domain/Person.java
package mg.rojo.domain;

import java.util.Calendar;
import java.util.Date;
import java.util.Objects;

/**
 * Created by dida on 11/22/16.
 */
public class Person {

    private String firstName;
    private String lastName;
    private Boolean gender;
    private Date dateOfBirth;
    private Company company;
    
	public Person(String firstName, String lastName, Boolean gender, Date dateOfBirth, Company company) {
		this.firstName = Objects.requireNonNull(firstName);
		this.lastName = Objects.requireNonNull(lastName);
		this.dateOfBirth = Objects.requireNonNull(dateOfBirth);
		this.gender = Objects.requireNonNull(gender);
		this.company = Objects.requireNonNull(company);
	}
	public String getFirstName() {
		return firstName;
	}
	public String getLastName() {
		return lastName;
	}
	public Boolean getGender() {
		return gender;
	}
	public Date getDateOfBirth() {
		return dateOfBirth;
	}
	public Company getCompany() {
		return company;
	}
	@Override
	public int hashCode() {
		final int prime = 31;
		int result = 1;
		result = prime * result + ((company == null) ? 0 : company.hashCode());
		result = prime * result + ((dateOfBirth == null) ? 0 : dateOfBirth.hashCode());
		result = prime * result + ((firstName == null) ? 0 : firstName.hashCode());
		result = prime * result + ((gender == null) ? 0 : gender.hashCode());
		result = prime * result + ((lastName == null) ? 0 : lastName.hashCode());
		return result;
	}
	@Override
	public boolean equals(Object obj) {
		if (this == obj)
			return true;
		if (obj == null)
			return false;
		if (getClass() != obj.getClass())
			return false;
		Person other = (Person) obj;
		if (company == null) {
			if (other.company != null)
				return false;
		} else if (!company.equals(other.company))
			return false;
		if (dateOfBirth == null) {
			if (other.dateOfBirth != null)
				return false;
		} else if (!dateOfBirth.equals(other.dateOfBirth))
			return false;
		if (firstName == null) {
			if (other.firstName != null)
				return false;
		} else if (!firstName.equals(other.firstName))
			return false;
		if (gender == null) {
			if (other.gender != null)
				return false;
		} else if (!gender.equals(other.gender))
			return false;
		if (lastName == null) {
			if (other.lastName != null)
				return false;
		} else if (!lastName.equals(other.lastName))
			return false;
		return true;
	}
	
    
}
